+++ crates/polars-python/Cargo.toml
@@ -28,7 +28,7 @@ bytemuck = { workspace = true }
 ciborium = { workspace = true }
 either = { workspace = true }
 itoa = { workspace = true }
-libc = "0.2"
+libc = { workspace = true }
 ndarray = { workspace = true }
 num-traits = { workspace = true }
 # TODO: Pin to released version once NumPy 2.0 support is merged
@@ -110,16 +110,6 @@ features = [
 
 [build-dependencies]
 version_check = { workspace = true }
-
-[target.'cfg(all(any(not(target_family = "unix"), allocator = "mimalloc"), not(allocator = "default")))'.dependencies]
-mimalloc = { version = "0.1", default-features = false }
-
-# Feature background_threads is unsupported on MacOS (https://github.com/jemalloc/jemalloc/issues/843).
-[target.'cfg(all(target_family = "unix", not(target_os = "macos"), not(allocator = "mimalloc"), not(allocator = "default")))'.dependencies]
-jemallocator = { version = "0.5", features = ["disable_initial_exec_tls", "background_threads"] }
-
-[target.'cfg(all(target_family = "unix", target_os = "macos", not(allocator = "mimalloc"), not(allocator = "default")))'.dependencies]
-jemallocator = { version = "0.5", features = ["disable_initial_exec_tls"] }
 
 [features]
 # Features below are only there to enable building a slim binary during development.
