+++ crates/polars-python/build.rs
@@ -1,5 +1,4 @@
 fn main() {
-    println!("cargo::rustc-check-cfg=cfg(allocator, values(\"default\", \"mimalloc\"))");
     println!("cargo:rerun-if-changed=build.rs");
     let channel = version_check::Channel::read().unwrap();
     if channel.is_nightly() {
